refactor: move verifyRefreshTkn to auth utils
