+++ server/src/features/auth/utils/index.ts
@@ -2,3 +2,4 @@ export * from "./create-jwt.util";
 export * from "./create-payload.util";
 export * from "./create-tokens.util";
 export * from "./get-refresh-config.util";
+export * from "./verify-refresh-tkn";
